+++ src/idunno.AtProto/Handle.cs
@@ -25,7 +25,12 @@ namespace idunno.AtProto
         [DebuggerBrowsable(DebuggerBrowsableState.Never)]
         private const int MaximumLength = 253;
 
-        [GeneratedRegex(@"^([a-zA-Z0-9]([a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?\.)+[a-zA-Z]([a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?$", RegexOptions.None, 5000)]
+        /// <summary>
+        /// A regular expression suitable for use when validating a handle.
+        /// </summary>
+        public const string ValidationRegex = (@"^([a-zA-Z0-9]([a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?\.)+[a-zA-Z]([a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?$";
+
+        [GeneratedRegex(ValidationRegex, RegexOptions.None, 5000)]
         private static partial Regex s_validate();
 
         [SuppressMessage("Globalization", "CA1308:Normalize strings to uppercase", Justification = "AT Proto standards normalize to lower case.")]